add html_render

+++ fastapi/sql2.py
@@ -1,8 +1,10 @@
 from fastapi import APIRouter
+from fastapi.responses import HTMLResponse
+from jinja2 import Template, Environment, FileSystemLoader
 from pydantic import BaseModel
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, DeclarativeBase
-from sqlalchemy.schema import Column
+from sqlalchemy.orm import sessionmaker, DeclarativeBase, relationship
+from sqlalchemy.schema import Column, ForeignKey
 from sqlalchemy import Integer, String
 
 router = APIRouter(prefix="/sql2")
@@ -31,6 +33,13 @@ class CityInfo(BaseModel):
     population: int
 
 
+class Person(Base):
+    __tablename__ = "persons"
+    id = Column(Integer, primary_key=True)
+    name = Column(String)
+    city_id = Column(Integer, ForeignKey("cities.id"))
+
+
 Base.metadata.create_all(engine)
 
 @router.get("/test/")
@@ -46,11 +55,20 @@ def test2():
     city_a = session.query(City).get(1)
     city_a.name = "fuga"
     session.commit()
-    return get_all_users()
+    return get_all_cities_json()
+
+
+@router.get("/test3/")
+def test3():
+    person = Person(name="taro", city_id=3)
+    print(person)
+    session.add(person)
+    session.commit()
+    return
 
 
 @router.get("/all/")
-def get_all_users():
+def get_all_cities_json():
     response = session.query(City).all()
     return response
 
@@ -62,20 +80,29 @@ def delete_all():
 
 
 @router.delete("/{id}/")
-def delete_user(id: int):
+def delete_city(id: int):
     user_a = session.query(City).get(id)
-    session.delete(user_a)
-    session.commit()
-    return get_all_users()
+    if user_a:
+        session.delete(user_a)
+        session.commit()
+    return
 
 
 @router.post("/add/")
-def add_user(cityinfo: CityInfo):
+def add_city(cityinfo: CityInfo):
     session.add(City(name=cityinfo.name, population=cityinfo.population))
     session.commit()
     return
 
+env = Environment(loader=FileSystemLoader('./', encoding='utf8'))
+tmpl = env.get_template('sql2_html_template.j2')
 
-@router.get("/{id}")
-def get_user(id: int):
+@router.get("/city_list/", response_class=HTMLResponse)
+def get_all_cities_html():
+    return tmpl.render(items = get_all_cities_json())
+
+
+@router.get("/{id}/")
+def get_city(id: int):
     return session.query(City).get(id)
+
